Split up mullvad_update::version module
