+++ installer-downloader/src/environment.rs
@@ -1,5 +1,3 @@
-use mullvad_update::version::VersionArchitecture;
-
 /// The environment consists of globals and/or constants which need to be computed at runtime.
 pub struct Environment {
     pub architecture: Architecture,
@@ -22,15 +20,15 @@ impl Environment {
 
     /// Try to map the host's CPU architecture to one of the CPU architectures the Mullvad VPN app
     /// supports.
-    fn get_arch() -> Result<VersionArchitecture, Error> {
+    fn get_arch() -> Result<Architecture, Error> {
         let arch = talpid_platform_metadata::get_native_arch()
             .inspect_err(|err| log::debug!("{err}"))
             .map_err(|_| Error::Arch)?
             .ok_or(Error::Arch)?;
 
         let arch = match arch {
-            talpid_platform_metadata::Architecture::X86 => VersionArchitecture::X86,
-            talpid_platform_metadata::Architecture::Arm64 => VersionArchitecture::Arm64,
+            talpid_platform_metadata::Architecture::X86 => Architecture::X86,
+            talpid_platform_metadata::Architecture::Arm64 => Architecture::Arm64,
         };
 
         Ok(arch)
